requirements.txt
streamlit
pandas
sqlalchemy
sqlite3
bcrypt
pyjwt
requests
streamlit_lottie
